CH Refactor: Rename folder
